continue working on chapter 2, IR almost done, still need to modify code_gen to use IR

+++ node.h
@@ -3,8 +3,8 @@
 
 struct node
 {
-    int   dataSize;
-    void* data;
+    size_t       dataSize;
+    void*        data;
     struct node* flink;
     struct node* blink;
 };
